pyproject.toml
[tool.poetry]
name = "django-aws-config"
version = "0.0.0"
description = ""
authors = ["TreyWW"]
#readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django = "^5.0"
requests = "^2.31.0"
types-requests = "^2.31.0.20240311"
typing_extensions = "^4.10.0"
bump-my-version = "^0.19.3"
pydantic = "^2.7.1"
django-storages = "^1.14.2"
boto3 = "^1.34.98"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
coverage = "^7.4.4"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.mkdocs.dependencies]
mkdocs = "^1.5.3"
mkdocs-material = "^9.5.15"
mike = "2.0.0"

[tool.poetry.group.mkdocs]
optional = true

[tool.black]
line-length = 140

[build-system]
requires = ['setuptools>=40.8.0', 'django>=5.0']
build-backend = 'setuptools.build_meta'


[project]
name = "django-aws-config"
description = "A django package aimed to simplify payment integration into your application"
readme = "README.md"
authors = [
    {name = "Trey", email="trey@strelix.org"}
]
requires-python = ">=3.10"

license = {file = "LICENSE"}

classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]

dynamic = [
    "version"
]

dependencies = [
    'django>=5.0',
    'pydantic>=2.7.1',
    'requests>=2.31.0',
    'django-storages>=1.14.2',
    'boto3>=1.34.98'
]

[tool.setuptools.dynamic]
version = {attr = "aws_config.__version__"}

[project.urls]
Homepage = "https://github.com/Strelix/django-aws-config"
Issues = "https://github.com/Strelix/django-aws-config/issues"

[tool.bumpversion]
current_version = "0.0.0"
commit = true
commit_args = "-s"
tag = true
tag_name = "v{new_version}"
sign_tags = true
allow_dirty = true
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)(\\.(?P<dev>post)\\d+\\.dev\\d+)?"
serialize = [
    "{major}.{minor}.{patch}.{dev}{$PR_NUMBER}.dev{distance_to_latest_tag}",
    "{major}.{minor}.{patch}"
]

message = "Version updated from {current_version} to {new_version}"

[[tool.bumpversion.files]]
filename = "aws_config/__init__.py"

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'